Refactor basic-no-plopfile
